rework formant and interval method
The original goal of my 2020 rewrite of this script was to first modify a new Sound object so that it consisted only of the voiced segments from the original object, then calculate the formant traces on this new Sound. However, this results in problematic formant values at each point where the window used for formant calculation spans the boundary where an unvoiced segment was removed. To address this, the formant calculation is now done on the original file, but time points for each voiced segment are output, as well. These can be used in Pandas to filter the formant traces and remove unvoiced sections there. (It should be possible to perform this same check in Praat, also, but by passing the whole formant traces to Pandas, smoothing can be done there prior to the removal of unvoiced segments.)