formant collection code as a function
The function in this file uses Praat to collect continuous formant data from an audio file, smooths this data, then limits the formant data to just those intervals for which voicing is detected. It returns the formant data as a Pandas DataFrame, making it easier to perform other operations with the data.